pyproject.toml
[tool.poetry]
name = "lazybot"
version = "0.1.0"
description = ""
authors = ["lazyaf7 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
python-dotenv = "^0.17.0"
replit = "^3.0.0"
flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
